move pkg/profiles to pkg/services/profiles
